Move AfkCheckTask to task folder
